integrated rpd-boardgame for the map, futzed with surfaces and blitting for ages.
Had some problem with blitting multiple surfaces together, was trying to have a sandwich:

 1. game board
 1. game pieces
 1. mouse effects
 1. chrome

With chrome on top, but I had trouble with alpha blending. I wanted to basically replace a segment
of the chrome surface with a new alpha fill, but it would always accumulate; instead fo cutting
out the rectangle and adding a new one with my desired alpha, it was painting on top, accumulating alpha
until it was opaque. I think fill-surface is the one I want, but there's still some odd effects
where fill-surface seems to be effecting a draw-string call later.